Move helpers to /lib
