+++ lib/escapeHTML.spec.ts
