+++ lib/escapeHTML.ts
