+++ lib/escapeRegExp.ts
